Load Mason immediately, other LSPs on lazy event.
Mason docs state that it's not recommended to lazy-load.

This was causing an issue where the LSP would not attach when running
`nvim <file>`, but would when opening nvim and then opening a file
(e.g., with Telescope or `:e`).

Now, Mason gets loaded immediately, and all other LSP-related stuff gets
lazy-loaded on BufReadPre and BufNew or whatever they're called.

This fixes the small issue described above.